update: remove useless abortController
